Crate safe_vault

SAFE Vault provides the interface to SAFE routing. The resulting executable is the Vault node for the SAFE network.

Macros

eval_result!
This macro is intended to be used in all cases where we unwrap() a result to delebrately panic in case of error - eg., in test-cases. Such unwraps don't give a precise point of failure in our code and instead indicate some line number in core library. This macro will provide a precise point of failure and will decorate the failure for easy viewing.

Functions

main
Runs a SAFE Network vault